requirements.txt
dash>=2.18
gunicorn
numpy
scipy
astro-tiptop==1.3.33